enhance journal transaction update coa balance

+++ Com.Danliris.Service.Finance.Accounting.Lib/ViewModels/JournalTransaction/GeneralLedgerReportViewModel.cs
@@ -8,18 +8,18 @@ namespace Com.Danliris.Service.Finance.Accounting.Lib.ViewModels.JournalTransact
     {
         public DateTimeOffset Date { get; set; }
         public string Description { get; set; }
-        public double Debit { get; set; }
-        public double Credit { get; set; }
+        public decimal Debit { get; set; }
+        public decimal Credit { get; set; }
         public string COACode { get; set; }
-        public double RemainingBalance { get; set; }
+        public decimal RemainingBalance { get; set; }
     }
 
     public class GeneralLedgerWrapperReportViewModel
     {
         public string COACode { get; set; }
         public string COAName { get; set; }
-        public double Summary { get; set; }
-        public double InitialBalance { get; set; }
+        public decimal Summary { get; set; }
+        public decimal InitialBalance { get; set; }
         public List<GeneralLedgerReportViewModel> Items { get; set; }
     }
 }
